server/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
scikit-learn>=1.4.0
requests>=2.31.0
Pillow>=10.2.0
python-multipart>=0.0.6